Stopwatch file reorg for command separation
